+++ CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Add Minor cosmetic improvements on gallery README.
+- (DEV) Add `gulp post-build-docs` to improve documentation output.
+> This version replaces documentation online links with local links
+> to support offline testing.
 
 
 ## [1.4.0] -